Rename QuicTransportOptions
Rename QuicTransportConfiguration to QuicTransportOptions to better
match web platform conventions. Also change
server_certificate_fingerprints to serverCertificateFingerprints.

As a side effect of regenerating index.html, incorporate the
ReceiveStreams example from #114 into the visible standard.

As a side effect of using a different version of Bikeshed, has some CSS
changes and updates the links to the Streams Standard to the latest
version.

Fixes #120.